Refactor to use lib funtion
